+++ ChangeLog.md
@@ -5,11 +5,20 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+## [0.2.0] - 2024-10-25
+
+### Changed
+
+- Changed casing references from `BlueSky` to `bluesky`.
+- Updated `README.md`.
+- Changed the global `PDSHOST`variable to a module-scoped variable.
+
 ## [0.1.0] - 2024-10-25
 
 ### Added
 
 - initial files and module structure
 
-[Unreleased]: https://github.com/jdhitsolutions/PSBlueSky/compare/v0.1.0..HEAD
+[Unreleased]: https://github.com/jdhitsolutions/PSBluesky/compare/v0.2.0..HEAD
+[0.2.0]: https://github.com/jdhitsolutions/PSBluesky/compare/v0.1.0..v0.2.0
 [0.1.0]: